pyproject.toml
[project]
name = "free-soph-fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "aiohttp-socks>=0.10.1",
    "fake-useragent>=2.2.0",
    "fastapi>=0.115.12",
    "httptools>=0.6.4",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "uvicorn[standard]>=0.34.2",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = ["pytest>=8.3.5"]

[project.scripts]
openai_chat_proxy = "main:main"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
